Added support for peptide level precursor charge, note: PRIDE Chart is not yet working.

+++ pride-inspector/branches/development/src/main/java/uk/ac/ebi/pride/gui/task/impl/AbstractConnectPrideTask.java
@@ -102,7 +102,7 @@ public abstract class AbstractConnectPrideTask extends TaskAdapter<List<Map<Stri
             boutStream = new BufferedOutputStream(outStream, BUFFER_SIZE);
             byte data[] = new byte[BUFFER_SIZE];
             int count;
-            int readCount = 0;
+            long readCount = 0;
             if (size != null) {
                 setProgress(1);
             }
